refactor: rename `app` package to `hd_active`
